Add active_sale_prices association
It adds `active_sale_prices` association to `Spree::Price` in order to
fetch active sale prices so to take advantage of Rails associations
features (primarily cache).

The association addition requires us to reload variant's default price
when we perform sale operations on their prices because otherwise it'd
go out-of-sync with them.